requirements.txt
Flask==2.0.1
requests==2.26.0
gunicorn==20.1.0
python-dotenv==0.19.0 